+++ src/index.ts
@@ -3,7 +3,7 @@
  * @Author       : Yp Z
  * @Date         : 2023-08-14 18:01:15
  * @FilePath     : /src/index.ts
- * @LastEditTime : 2023-11-28 14:01:53
+ * @LastEditTime : 2023-12-05 20:04:21
  * @Description  : 
  */
 import {
@@ -339,6 +339,16 @@ export default class RunJsPlugin extends Plugin {
      * @param code string, 代码字符串
      */
     public async runJsCode(code: string, codeBlock?: Block) {
+        let func = new Function(
+            'siyuan', 'client', 'api', 'plugin', 'thisBlock',
+            code
+        );
+        return func(siyuan, client, api, this, codeBlock);
+    }
+
+    public runJsCodeAsync = this.runJsCode;
+
+    public runJsCodeSync(code: string, codeBlock?: Block) {
         let func = new Function(
             'siyuan', 'client', 'api', 'plugin', 'thisBlock',
             code